% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_env_uncertainty.R
\name{assess_env_uncertainty}
\alias{assess_env_uncertainty}
\title{Compute variation of environmental conditions.}
\usage{
assess_env_uncertainty(x, y, by = 1000, scale = 1000)
}
\arguments{
\item{x}{ee$Image or ee$ImageCollection objects with a single band.}

\item{y}{ee$Geometry$*, ee$Feature, ee$FeatureCollection or sf objects.}

\item{by}{Numerical input. Numbers of features.}

\item{scale}{A nominal scale in meters of the Image projection to work in. By default 1000.}
}
\value{
The function returns a sf object with column called "elevation" with standard deviation for each feature.
}
\description{
Given a grid system and environmental data this function compute the standard deviation for each feature of the grid system.
}
\note{
The functions with prefix "ee_" is based on rgee package for interacting with Google Earth Engine (GEE). To run these functions or anything related to rgee/GEE, users must have:
\itemize{
\item{Google account with Earth Engine activated}
\item{Python >= v3.5}
\item{EarthEngine Python API (Python package)}
}
If the strict dependencies are not installed, rgee just will not work. It highly recommended seeing the installations and activation instructions in \href{https://cran.r-project.org/web/packages/rgee/vignettes/rgee01.html#6_Installation}{rgee} documentation
}
\examples{
\dontrun{
# As this function is based on rgee package, the following commands must be executed:

# Create virtual environmental at local machine to install the necessary dependencies
rgee::ee_install()

# Initialize
rgee::ee_Initialize()

# Load ImageCollection of interest i.e, an environmental layer

nasadem<- rgee::ee$Image('NASA/NASADEM_HGT/001')$select('elevation')

# Hypothetical grid system
lat_lon_grid <- structure(list(ID = 758432:758443,
                              lat = c(-14.875, -14.875, -14.625, -14.625, -14.875, -14.875, -14.625, -14.625, -14.375, -14.375, -14.125, -14.125),
                              lon = c(-42.875, -42.625, -42.625, -42.875, -42.375, -42.125, -42.125, -42.375, -42.375, -42.125, -42.125, -42.375)),
                         class = "data.frame", row.names = c(NA, -12L))

grid_to_raster <- raster::rasterFromXYZ(lat_lon_grid [, c('lon', 'lat', 'ID')], crs = '+proj=longlat +datum=WGS84 +no_defs')

grid <- raster::rasterToPolygons(grid_to_raster, fun=NULL, na.rm=TRUE, dissolve=FALSE) |>
 sf::st_as_sf() |>
 rgee::sf_as_ee()

# Plot
rgee::Map$addLayer(grid)

# Compute standard deviation
std_dev <- GridDER::assess_env_uncertainty(x= nasadem, y= grid)
}

}
\seealso{
\href{https://csaybar.github.io/rgee-examples/}{rgee examples} and \href{https://cran.r-project.org/web/packages/rgee/vignettes/rgee01.html}{rgee cran documentation}.
}
\author{
Tainá Rocha
}
