% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_sp_uncertainty.R
\name{assess_sp_uncertainty}
\alias{assess_sp_uncertainty}
\title{@title  Assesses the spatial uncertainty of a grid system by calculating half of the diagonal of a grid cell}
\usage{
assess_sp_uncertainty(res_x, res_y)
}
\arguments{
\item{res_x}{Resolution of a grid system along horizontal axis}

\item{res_y}{Resolution of a grid system along vertical axis}
}
\description{
@title  Assesses the spatial uncertainty of a grid system by calculating half of the diagonal of a grid cell
}
