% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_angle.R
\name{cal_angle}
\alias{cal_angle}
\title{Cal Angle function}
\usage{
cal_angle(M, N)
}
\arguments{
\item{M}{Coordinate of the reference point}

\item{N}{Coordinate of the reference point}
}
\description{
This function calculates the angle between two points. It uses one point as the reference center. 0 degree means top, 90 degree means right, 180 degree means below, and 270 degree means left.
}
\author{
Xiao Feng
}
