% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_distance_to_grid.R
\name{cal_distance_to_grid}
\alias{cal_distance_to_grid}
\title{Compute nearest distance}
\usage{
cal_distance_to_grid(
  input_occ = NULL,
  country = NULL,
  gridID = NULL,
  grid_metadata = NULL,
  flag_over_XPer = 10,
  flag_rm_Large_outlier = TRUE,
  flag_user_threshold = NULL
)
}
\arguments{
\item{input_occ}{Input occurrences (shapefile or .txt/.csv data with decimalLongitude & decimalLatitude).}

\item{country}{Character vector. Country name; can be used to subset known grid systems (G sheet, or a dataframe from user, in which case it should have a similar structure as G sheet).}

\item{gridID}{Character vector. Grid ID to be used in the calculation (gridID from G sheet or user provided dataframe).}

\item{grid_metadata}{A dataframe. If null, G sheet is loaded; otherwise, user should provide a table that looks like G sheet.}

\item{flag_over_XPer}{Numerical input. If a relative distance is >0.1 or 10\% of max distance (half of the diagonal line), then make it NA (meaning not from a grid).}

\item{flag_rm_Large_outlier}{Logical. Give a pool of distances, the remove outliers (based on R boxplot algorithm)}

\item{flag_user_threshold}{Numerical input. A number between 0 and 100; representing a threshold of relative distance (actual distance/(half of the diagonal line), beyond of which an occurrence is excluded.}
}
\value{
The funtion returns the most close grid system for each occurrence, based on abs distance, relative distance, or both distance.
}
\description{
This function compute  nearest distance  between a set of occurrences and a set of grid systems.
}
\examples{
\dontrun{
# Add example
}
}
