% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_emp_G_distance.R
\name{cal_emp_G_distance}
\alias{cal_emp_G_distance}
\title{Compute the distance between a grid system and occurrences}
\usage{
cal_emp_G_distance(
  input_grid,
  input_occ_grid,
  input_occ_random,
  flag_degree = FALSE,
  flag_plot = FALSE
)
}
\arguments{
\item{input_grid}{Character vector. Path of a grid system (shapefile).}

\item{input_occ_grid}{Character vector. Path of gridded occurrences (shapefile).}

\item{input_occ_random}{Character vector. Path of non-gridded occurrences (inaturalist is used by default).}

\item{flag_degree}{Logical. Parameter passed to raster::pointDistance.}

\item{flag_plot}{Logical. If TRUE, coordinates (based on the crs of the grid system) should be in degrees; else they should represent planar ('Euclidean') space (e.g. units of meters). Default is FALSE.}
}
\value{

}
\description{
The function compute the
}
\examples{
\dontrun{
# Will change. Not run!
test1 <- cal_emp_G_distance(input_grid = "data/2_clean_grid/grid_ID_2_polygon_masked.shp",
input_occ_grid = "data/france_datasetID10/0068190-210914110416597.csv",
input_occ_random = "data/iNaturalist_20211205/0071565-210914110416597.csv")
}
}
