% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_demoGrid.R
\name{download_demoGrid}
\alias{download_demoGrid}
\title{Download grid systems}
\usage{
download_demoGrid(downloadNew = F)
}
\arguments{
\item{downloadNew}{Logical. TRUE to download the zip file containing all grid systems}
}
\description{
This functions load from Zenodo the metadata with all grid systems compiled.
}
\author{
Xiao Feng
}
