% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_stdDev.R
\name{ee_stdDev_stat}
\alias{ee_stdDev_stat}
\title{Compute standard deviation}
\usage{
ee_stdDev_stat(x, y, by = 1000, scale = 1000)
}
\arguments{
\item{x}{ee$Image or ee$ImageCollection objects with a single band.}

\item{y}{ee$Geometry$*, ee$Feature, ee$FeatureCollection or sf objects.}

\item{by}{Numerical input. Numbers of features.}

\item{scale}{A nominal scale in meters of the Image projection to work in. By default 1000.}
}
\value{

}
\description{
Given a grid system and environmental data this function compute the standard deviation for each feature of the grid system.
}
\examples{
\dontrun{
# Read the grid system of interest
gridID_26 <- rgdal::readOGR("./data/South_Africa/grid_id_26/v2_grid_id_26.shp") |>
sf::st_as_sf()

# Countries Map Base
base_map <- rgdal::readOGR("./data/0_basemap/ne_10m_admin_0_countries.shp") |>
sf::st_as_sf()

#Select a Region of Interest
South_Africa <- base_map |>
dplyr::filter(ADMIN=="South Africa") |>
sf::st_intersection(gridID_26) |>
dplyr::select(ADMIN, geometry
}
}
