% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_CRS_extent.R
\name{find_CRS_extent}
\alias{find_CRS_extent}
\title{Find CRS}
\usage{
find_CRS_extent(crs_num = "2154")
}
\arguments{
\item{crs_num}{}
}
\value{

}
\description{
This function is used by generateGRID function, for auto extracting information from epsg website
}
