% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist_freq.R
\name{get_dist_freq}
\alias{get_dist_freq}
\title{Find the Distance}
\usage{
get_dist_freq(input_v, round_num = 0)
}
\arguments{
\item{input_v}{Vector? X or Y coordinates}

\item{round_num}{Numeric}
}
\description{
Find the distance of x(or y), then return a frequency table
}
\author{
Xiao Feng
}
