% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_adjustment.R
\name{grid_adjustment}
\alias{grid_adjustment}
\title{Adjust the Grid System}
\usage{
grid_adjustment(input_grid, input_occ_grid)
}
\arguments{
\item{input_grid}{Character vector. A string with path of input grid system.}

\item{input_occ_grid}{Character vector. A string with path of occurrences data.}
}
\value{
The function returns the grid adjusted (shapefile).
}
\description{
Given a grid system and occurrences, this function try to shift the grid system, in order to minimize the distance between the grid system and input occurrences i.e. doing some small adjustment of the grid system.
}
\author{
Xiao Feng
}
