% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_matching.R
\name{grid_matching}
\alias{grid_matching}
\title{Compute nearest distance}
\usage{
grid_matching(
  input_occ,
  input_grid = NULL,
  country = NULL,
  grid_metadata = NULL,
  flag_relativeTHD = NULL,
  flag_rm_Large_outlier = FALSE,
  flag_absoluteTHD = NULL
)
}
\arguments{
\item{input_occ}{Input occurrences (shapefile or .txt/.csv data with decimalLongitude & decimalLatitude).}

\item{input_grid}{A list of spatial polygons, each represent one grid system.}

\item{country}{Character vector. A string or a vector of strings of countries names; can be used to subset known grid systems (G sheet, or a dataframe from user, in which case it should have a similar structure as G sheet).}

\item{grid_metadata}{A dataframe, if null, G sheet is loaded; otherwise, user should provide a table that looks like G sheet.}

\item{flag_relativeTHD}{Numerical. 0.1 if a relative distance is >0.1 or 10\% of max distance (half of the diagonal line), then make it NA (meaning not from a grid).}

\item{flag_rm_Large_outlier}{Give a pool of distances, then remove outliers (based on R boxplot algorithm).}

\item{flag_absoluteTHD}{Numerical. An absolute distance, beyond of which an occurrence is excluded.}

\item{gridID}{Character vector. A string or a vector of strings; grid ID to be used in the calculation (gridID from G sheet or user provied dataframe).}
}
\description{
This function calculate nearest distance between a set of occurrences and a set of grid systems.
}
\author{
Xiao Feng
}
