% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_crs.R
\name{infer_crs}
\alias{infer_crs}
\title{Infer Crs}
\usage{
infer_crs(
  occ_path,
  truth_crs_num = NA,
  flag_saveTemp = TRUE,
  flag_newCal = TRUE,
  temp_path = "data/3_infer_grid_crs/prj_occ_temp_v2/",
  cup_num = 10,
  flag_debug = -1
)
}
\arguments{
\item{occ_path}{Character. A string, path of input occurrences, should have decimalLongitude & decimalLatitude.}

\item{truth_crs_num}{Character. A string, epsg code of the true crs if known.}

\item{flag_saveTemp}{Logical. If "True", save the middle output.}

\item{flag_newCal}{Logical. If "True" do the distance calculation again.}

\item{temp_path}{Character. Path to save.}

\item{flag_debug}{}
}
\value{

}
\description{
This function will prj occ to a list of ~6600 crs, find  nearest 4 points, find the angle to the 4 points, then take reminder \%\%90, the smaller the better also find the distance to the 4 points, then all the distance is standardized to 0 mean and 1sd  then do a histogram, and calculate the freq of each bin,  pick the bin with highest freq ideally, the true crs will lead to the highest freq in that bin.
}
\note{
Using 15 cores, it took ~10-20 minutes to finish this workflow.
}
\author{
Xiao Feng
}
