% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_extent.R
\name{infer_extent}
\alias{infer_extent}
\title{Infer extent}
\usage{
infer_extent(
  method = "",
  country = NULL,
  occ = NULL,
  crs_grid = NULL,
  flag_adjust_by_res = FALSE,
  res_x = NULL,
  res_y = NULL,
  flag_adjust_origin = (method == "occ_extent"),
  flag_unit = "meter"
)
}
\arguments{
\item{method}{Character. A string, country_extent, crs_extent, occ_extent}

\item{country}{Character. A string of country}

\item{flag_adjust_by_res}{Logical.}

\item{flag_unit}{Character. A a string; default value is meter "meter". It also can "degree","minute","second"}
}
\description{
Return an extent in the order of xmin, xmax, ymin, ymax
}
\author{
Xiao Feng
}
