% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_origin.R
\name{infer_origin}
\alias{infer_origin}
\title{Infer Origin}
\usage{
infer_origin(
  input_occ,
  fun = min,
  flag_unit = "meter",
  filters = sort(c(10^(2:10), 10^(3:10)/2))
)
}
\arguments{
\item{flag_unit}{Character. Should be a string; default value is meter It also can "degree","minute","second"}

\item{filters}{}
}
\value{

}
\description{
Find the minimum and maximum of occurrence. Here the minimum and maximum should match the resolution of the grid
}
\author{
Xiao Feng
}
