% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_resolution.R
\name{infer_resolution}
\alias{infer_resolution}
\title{Infer Resolution}
\usage{
infer_resolution(input_coord, digits = 1, ...)
}
\arguments{
\item{input_coord}{Coordinates of spatial points}

\item{...}{Additional parameters}
}
\description{
Compute the resolution of the grid system
}
\author{
Xiao Feng
}
