% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_occ.R
\name{load_occ}
\alias{load_occ}
\title{Load occurrence records}
\usage{
load_occ(path, crs_num = NULL)
}
\arguments{
\item{path}{Character vector. Path of occurrences data, should have decimalLongitude & decimalLatitude.}

\item{crs_num}{Character. Crs of the output spatial data points.}
}
\value{
Unique records.
}
\description{
This function read a GBIF dataset, then return the unique coordinates in spatial format.
}
\examples{
\dontrun{
# Will change. Not run!
unique_records <- load_occ("data-raw/Netherlands/gridID_24/0047643-210914110416597.csv")
}
}
\author{
Xiao Feng
}
