% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_sp_uncertainty.R
\name{assess_sp_uncertainty}
\alias{assess_sp_uncertainty}
\title{Assess the spatial uncertainty}
\usage{
assess_sp_uncertainty(
  input_grid,
  input_occ_grid,
  input_occ_random,
  flag_degree = FALSE,
  flag_plot = FALSE
)
}
\arguments{
\item{input_grid}{Character vector. Path of a grid system (shapefile).}

\item{input_occ_grid}{Character vector. Path of gridded occurrences (shapefile).}

\item{input_occ_random}{Character vector. Path of non-gridded occurrences (inaturalist is used by default).}

\item{flag_degree}{Logical. Parameter passed to raster::pointDistance.}

\item{flag_plot}{Logical. If TRUE, coordinates (based on the crs of the grid system) should be in degrees; else they should represent planar ('Euclidean') space (e.g. units of meters). Default is FALSE.}
}
\description{
The function computes spatial uncertainty of a grid system by calculating the distance between grid centroids toward gridded coordinates and non-gridded coordinates (e.g. iNaturalist data).
}
\examples{
 \dontrun{

distance <- cal_emp_G_distance(input_grid = "data/2_clean_grid/grid_ID_2_polygon_masked.shp",
input_occ_grid = "data/france_datasetID10/0068190-210914110416597.csv",
input_occ_random = "data/iNaturalist_20211205/0071565-210914110416597.csv")
}
}
\author{
Xiao Feng
}
