% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_stdev.R
\name{ee_stdDev}
\alias{ee_stdDev}
\title{Standard deviation}
\usage{
ee_stdDev(x, y, by = 1000, scale = 1000)
}
\arguments{
\item{x}{ee$Image or ee$ImageCollection objects with a single band.}

\item{y}{ee$Geometry$*, ee$Feature, ee$FeatureCollection or sf objects.}

\item{by}{Numerical input. numbers of features}

\item{scale}{A nominal scale in meters of the Image projection to work in. By default 1000.}
}
\value{
a sf object
}
\description{
Standard deviation
}
