% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_grid.R
\name{generateGRID}
\alias{generateGRID}
\title{Generate a grid system}
\usage{
generateGRID(
  res_x = 10,
  res_y = 10,
  unit = "km",
  flag_crs = TRUE,
  extent_unit = NULL,
  input_extent = NULL,
  country = NULL,
  crs_num = NULL,
  crs_obj = NULL,
  flag_offset = NULL,
  flag_maskByCountry = FALSE,
  flag_buffer = 2,
  flag_loadCountryPolygon = TRUE
)
}
\arguments{
\item{res_x}{A numeric input of longitude resolution.}

\item{res_y}{A numeric input of latitude resolution.}

\item{unit}{Character vector of unite in minute ("minute") or kilometre ("km").}

\item{flag_crs}{Logical. Flag the CRS.}

\item{extent_unit}{Character vector. It can be "crs_web_extent" for... or "crs_countryPolygon" for..}

\item{country}{Character vector of countries. Ex.: "Germany"}

\item{crs_num}{Character of CRS code . Ex.: "4326"}

\item{flag_loadCountryPolygon}{}
}
\value{

}
\description{
This function is used to generate(simulate) a grid system based on metadata we have used our metadata to simulate some grid systems (2_simulate_grid.Rmd).
}
