% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_resolution.R
\name{infer_resolution}
\alias{infer_resolution}
\title{Infer Resolution}
\usage{
infer_resolution(input_coord, digits = 1, flag_unit = "meter", ...)
}
\arguments{
\item{...}{}
}
\description{
Infer Resolution
}
\author{
Xiao Feng
}
